FlagReset only when flaggroup feature active

+++ advancedregionmarket/src/main/java/net/alex9849/arm/regions/Region.java
@@ -695,7 +695,8 @@ public abstract class Region implements Saveable {
             i--;
         }
 
-        this.getRegion().deleteAllFlags();
+        if(FlagGroup.isFeatureEnabled())
+            this.getRegion().deleteAllFlags();
         if (this.isSubregion()) {
             AdvancedRegionMarket.getInstance().getWorldGuardInterface().removeFromRegionManager(this.getRegion(), this.getRegionworld());
             this.getParentRegion().getSubregions().remove(this);
